Clean up column, batch and stream objects
