+++ include/hipipe/core/stream.hpp
@@ -13,9 +13,12 @@
 
 /// \defgroup Stream Stream modifiers and data types.
 
+#include <hipipe/core/stream/batch_t.hpp>
+#include <hipipe/core/stream/column_t.hpp>
+#include <hipipe/core/stream/stream_t.hpp>
+
 #include <hipipe/core/stream/rebatch.hpp>
 #include <hipipe/core/stream/buffer.hpp>
-#include <hipipe/core/stream/column.hpp>
 #include <hipipe/core/stream/create.hpp>
 #include <hipipe/core/stream/drop.hpp>
 #include <hipipe/core/stream/filter.hpp>
