+++ include/hipipe/core/stream/unpack.hpp
@@ -10,7 +10,7 @@
 
 #pragma once
 
-#include <hipipe/core/stream/column.hpp>
+#include <hipipe/core/stream/stream_t.hpp>
 #include <hipipe/core/stream/template_arguments.hpp>
 #include <hipipe/core/utility/ndim.hpp>
 #include <hipipe/core/utility/tuple.hpp>
